backend/requirements.txt
# Phase 1 - Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Phase 1 Part 2 - Embedding, Vector DB, Clustering
chromadb>=0.4.0
sentence-transformers>=2.2.0
hdbscan>=0.8.33
numpy>=1.24.0

# Phase 1 Part 3 - Mistral extraction
mistralai>=1.0.0
python-dotenv>=1.0.0
